requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-mpv<=1.0.8
gpiozero>=1.6
apscheduler>=3.10
ujson>=5.8
python-multipart>=0.0.9
pyyaml>=6.0
